rename StartGameScreen to PlayGameScreen
